+++ app/src/main/java/com/dede/android_eggs/views/settings/compose/groups/AboutGroup.kt
@@ -1,19 +1,21 @@
 package com.dede.android_eggs.views.settings.compose.groups
 
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.requiredSize
-import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Download
 import androidx.compose.material.icons.rounded.Info
 import androidx.compose.material.icons.rounded.Policy
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.em
 import com.dede.android_eggs.R
 import com.dede.android_eggs.util.CustomTabsBrowser
 import com.dede.android_eggs.views.settings.compose.basic.ExpandOptionsPref
@@ -34,16 +36,16 @@ fun AboutGroup() {
         VersionOption()
         Option(
             leadingIcon = {
-                Image(
-                    painter = painterResource(R.drawable.ic_pgyer_logo),
-                    contentDescription = stringResource(R.string.label_beta),
-                    modifier = Modifier
-                        .size(24.dp)
-                        .requiredSize(33.dp)
+                Text(
+                    "β",
+                    fontSize = 5.2.em,
+                    fontStyle = FontStyle.Italic,
+                    fontWeight = FontWeight.SemiBold,
+                    textAlign = TextAlign.Center,
+                    modifier = Modifier.width(24.dp),
                 )
             },
             title = stringResource(R.string.label_beta),
-            desc = stringResource(R.string.url_beta),
             trailingContent = imageVectorIconBlock(imageVector = Icons.Rounded.Download),
             onClick = {
                 CustomTabsBrowser.launchUrl(context, R.string.url_beta)