+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
-### v3.0.1
+### v3.1.0-beta01
+
+- Add Beta version track
+- Fix Android M Easter Egg players color [#387](https://github.com/hushenghao/AndroidEasterEggs/issues/387)
+- Upgrade project dependencies
+
+### v3.0.1 (2024-09-26)
 
 - Fix Android 15 Easter Egg add flag crash on Android 8.1 [#384](https://github.com/hushenghao/AndroidEasterEggs/issues/384)
-- Fix Android M Easter Egg players color [#387](https://github.com/hushenghao/AndroidEasterEggs/issues/387)
 - Upgrade project dependencies
 
 ### v3.0.0 (2024-09-18)